# `sns.swarmplot` that scales better