@mwaskom 2017-03-25
Thanks, this is good work.

The best path forward for further optimization is to handle points in the "gutters" (maximum displacement from center) separately. With large Ns, this will be the majority of points, and if much of the processing can be skipped on points that will end up in the gutter, it will save a lot of time.

I've always felt that it isn't necessary to put too much effort into optimization in the swarming algorithm as the plot itself doesn't scale that well to large number of points (as most end up overlapping themselves in the gutter). But even when the plot isn't going to work, it's nice to be able to find out quickly!